S5 codys work

+++ server/db/monsterDb.ts
@@ -1,4 +1,5 @@
 import connection from './connection.js'
+import fs from 'fs/promises'
 
 import {
   NewMonster,
@@ -46,8 +47,21 @@ export async function addMonster(newMonsterData: NewMonster) {
 }
 
 export async function deleteMonsterById(id: number) {
-  const result = await connection('monsters').where({ id }).delete()
-  return result
+  try {
+    const monster = await connection('monsters').where({ id }).first()
+    if (!monster) {
+      throw new Error('Monster not found')
+    }
+    const imageUrl = monster.image_url
+
+    await fs.unlink(imageUrl)
+
+    const result = await connection('monsters').where({ id }).delete()
+    return result
+  } catch (error) {
+    console.error('Error on delete monster by id', error)
+    throw error
+  }
 }
 
 export async function editMonsterName(
